Update dependencies so this builds
As time marches on, so too do our dependencies.  In the client, webpack
needed to be updated to 3.0.0 due to a conflict with expose-loader.  It
may be able to be updated to 4.0.0, but that requires a webpack CLI
which I didn't feel like bothering with.

On the server side, there needed to be added a couple packages to make
the HTTP server work.  I don't know exactly what was up.